Extract config s.t. only in Runtime.
