+++ build-system/test-configs/forbidden-terms.js
@@ -631,7 +631,7 @@ const forbiddenTermsGlobal = {
       'build-system/tasks/build.js',
       'build-system/tasks/default-task.js',
       'build-system/tasks/dist.js',
-      'src/config.js',
+      'src/runtime-config.js',
       'src/core/window/window.extern.js',
       'src/experiments/index.js',
       'src/experiments/shame.extern.js',
@@ -967,7 +967,7 @@ const forbiddenTermsSrcInclusive = {
   '(cdn|3p)\\.ampproject\\.': {
     message:
       'The CDN domain should typically not be hardcoded in source ' +
-      'code. Use a property of urls from src/config.js instead.',
+      'code. Use a property of urls from src/runtime-config.js instead.',
     allowlist: [
       'ads/_a4a-config.js',
       'build-system/server/amp4test.js',
@@ -981,7 +981,7 @@ const forbiddenTermsSrcInclusive = {
       'build-system/tasks/performance/helpers.js',
       'src/3p-frame.js',
       'src/amp-story-player/amp-story-player-impl.js',
-      'src/config.js',
+      'src/runtime-config.js',
       'testing/local-amp-chrome-extension/background.js',
       'tools/experiments/experiments.js',
       'validator/js/engine/htmlparser-interface.js',
